prevPrime
Write the function prevPrime that receives an integer and returns the largest prime number that is smaller than the entered integer.
If there is not a prime number that is smaller than the entered integer, the function should return 0.
Use the function in a program that receives an integer from user and displays the prime numbers smaller than the entered integer.